refactor(Pause): Optimized the Pause Functionality
This commit optimizes the pausing functionality. It does this by eliminating a few key things. First, it eliminates the need to call GetComponent and AddListenener on every frame through the update function. Second, It removes the need to call GameObject.Find in start because GameObject.Find is very slow, as it has to scan the entire hierarchy ever time to find the object. Also, it can cause issues that are difficult to debug if you accidentally rename the object. Third, we eliminate the singleton structure as it is unnecessary. As no apis need to be exposed, there is no need for it.